dear imgui text input event

+++ plugin/box2d/impl/shapes/polygon.cpp
@@ -23,7 +23,7 @@ void Polygon::apply(b2Body* body, const sp<Size>& size, const BodyCreateInfo& cr
     b2Vec2* vecs = new b2Vec2[_vertices.size()];
     for(uint32_t i = 0; i < _vertices.size(); i ++)
         vecs[i].Set(_vertices[i].x() * size->width() / 2.0f, _vertices[i].y() * size->height() / 2.0f);
-    shape.Set(vecs, _vertices.size());
+    shape.Set(vecs, static_cast<int32_t>(_vertices.size()));
     delete[] vecs;
 
     b2FixtureDef fixtureDef = createInfo.toFixtureDef(&shape);
